# More type erasure for AnySegmentIterable to reduce compile times in Debug build