--- a/cmake/FindFS.cmake
+++ b/cmake/FindFS.cmake
@@ -4,7 +4,7 @@
 #  FILESYSTEM_FOUND    : True if found.
 
 if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU" OR UNIX AND NOT APPLE)
-    FIND_LIBRARY(FILESYSTEM_LIBRARY NAME libstdc++fs.a HINTS $ENV{HOME}/local/lib64 $ENV{HOME}/local/lib /usr/local/lib64 /usr/local/lib /opt/local/lib64 /opt/local/lib /usr/lib64 /usr/lib /usr/lib/gcc/*/*/ /usr/local/Cellar/gcc/*/lib/gcc/*/)
+    FIND_LIBRARY(FILESYSTEM_LIBRARY NAME libstdc++fs.a HINTS $ENV{HOME}/local/lib64 $ENV{HOME}/local/lib /usr/local/lib64 /usr/local/lib /opt/local/lib64 /opt/local/lib /usr/lib64 /usr/lib /usr/lib/gcc/*/8/ /usr/local/Cellar/gcc/*/lib/gcc/*/)
 elseif ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang")
     FIND_LIBRARY(FILESYSTEM_LIBRARY NAMES libc++fs.a libc++experimental.a HINTS $ENV{HOME}/local/lib64 $ENV{HOME}/local/lib /usr/local/lib64 /usr/local/lib /opt/local/lib64 /opt/local/lib /usr/lib64 /usr/lib /usr/local/Cellar/llvm/*/lib)
 endif()