--- a/src/bin/playground.cpp
+++ b/src/bin/playground.cpp
@@ -8,4 +8,4 @@ int main() {
   std::cout << "Hello world!!" << std::endl;
   return 0;
 }
- "this b"
+