--- a/src/lib/operators/jit_operator/operators/jit_read_tuples.cpp
+++ b/src/lib/operators/jit_operator/operators/jit_read_tuples.cpp
@@ -2,7 +2,7 @@
 
 #include "../jit_types.hpp"
 #include "resolve_type.hpp"
-#include "storage/create_iterable_from_segment.hpp"
+#include "storage/segment_iteration.hpp"
 
 namespace opossum {
 
@@ -73,18 +73,18 @@ void JitReadTuples::before_chunk(const Table& in_table, const Chunk& in_chunk, J
     const auto column_id = input_column.column_id;
     const auto segment = in_chunk.get_segment(column_id);
     const auto is_nullable = in_table.column_is_nullable(column_id);
-    resolve_data_and_segment_type(*segment, [&](auto type, auto& typed_segment) {
-      using Type = typename decltype(type)::type;
-      create_iterable_from_segment<Type>(typed_segment).with_iterators([&](auto it, auto end) {
-        using IteratorType = decltype(it);
-        if (is_nullable) {
-          context.inputs.push_back(
-              std::make_shared<JitSegmentReader<IteratorType, Type, true>>(it, input_column.tuple_value));
-        } else {
-          context.inputs.push_back(
-              std::make_shared<JitSegmentReader<IteratorType, Type, false>>(it, input_column.tuple_value));
-        }
-      });
+
+    segment_with_iterators(*segment, [&](auto it, const auto end) {
+      using Type = typename decltype(it)::ValueType;
+      using IteratorType = decltype(it);
+
+      if (is_nullable) {
+        context.inputs.push_back(
+        std::make_shared<JitSegmentReader<IteratorType, Type, true>>(it, input_column.tuple_value));
+      } else {
+        context.inputs.push_back(
+        std::make_shared<JitSegmentReader<IteratorType, Type, false>>(it, input_column.tuple_value));
+      }
     });
   }
 }