--- a/src/lib/operators/jit_operator/operators/jit_read_tuples.cpp
+++ b/src/lib/operators/jit_operator/operators/jit_read_tuples.cpp
@@ -74,18 +74,22 @@ void JitReadTuples::before_chunk(const Table& in_table, const Chunk& in_chunk, J
     const auto segment = in_chunk.get_segment(column_id);
     const auto is_nullable = in_table.column_is_nullable(column_id);
 
+      if (is_nullable) {
     segment_with_iterators(*segment, [&](auto it, const auto end) {
       using Type = typename decltype(it)::ValueType;
       using IteratorType = decltype(it);
 
-      if (is_nullable) {
         context.inputs.push_back(
             std::make_shared<JitSegmentReader<IteratorType, Type, true>>(it, input_column.tuple_value));
-      } else {
+    });
+    } else {
+    segment_with_iterators(*segment, [&](auto it, const auto end) {
+      using Type = typename decltype(it)::ValueType;
+      using IteratorType = decltype(it);
         context.inputs.push_back(
             std::make_shared<JitSegmentReader<IteratorType, Type, false>>(it, input_column.tuple_value));
-      }
     });
+    }
   }
 }