--- a/src/lib/operators/join_hash/join_hash_steps.hpp
+++ b/src/lib/operators/join_hash/join_hash_steps.hpp
@@ -9,6 +9,7 @@
 #include "scheduler/current_scheduler.hpp"
 #include "scheduler/job_task.hpp"
 #include "storage/create_iterable_from_segment.hpp"
+#include "storage/segment_iteration.hpp"
 #include "type_cast.hpp"
 #include "type_comparison.hpp"
 #include "uninitialized_vector.hpp"
@@ -138,11 +139,14 @@ RadixContainer<T> materialize_input(const std::shared_ptr<const Table>& in_table
       // prepare histogram
       auto histogram = std::vector<size_t>(num_partitions);
 
-      resolve_segment_type<T>(*segment, [&, chunk_id](auto& typed_segment) {
-        auto reference_chunk_offset = ChunkOffset{0};
-        auto iterable = create_iterable_from_segment<T>(typed_segment);
+      auto reference_chunk_offset = ChunkOffset{0};
 
-        iterable.for_each([&, chunk_id](const auto& value) {
+      segment_with_iterators<T>(*segment, [&](auto it, const auto end) {
+        using IterableType = typename decltype(it)::IterableType;
+
+        while (it != end) {
+          const auto& value = *(it++);
+
           if (!value.is_null() || consider_null_values) {
             const Hash hashed_value = hash_function(type_cast<HashedType>(value.value()));
 
@@ -151,7 +155,7 @@ RadixContainer<T> materialize_input(const std::shared_ptr<const Table>& in_table
             Instead, we use the index in the ReferenceSegment itself. This way we can later correctly dereference
             values from different inputs (important for Multi Joins).
             */
-            if constexpr (std::is_same_v<std::decay<decltype(typed_segment)>, ReferenceSegment>) {
+            if constexpr (std::is_same_v<IterableType, ReferenceSegmentIterable<T>>) {
               *(output_iterator++) = PartitionedElement<T>{RowID{chunk_id, reference_chunk_offset}, value.value()};
             } else {
               *(output_iterator++) = PartitionedElement<T>{RowID{chunk_id, value.chunk_offset()}, value.value()};
@@ -169,10 +173,10 @@ RadixContainer<T> materialize_input(const std::shared_ptr<const Table>& in_table
             ++null_value_bitvector_iterator;
           }
           // reference_chunk_offset is only used for ReferenceSegments
-          if constexpr (std::is_same_v<std::decay<decltype(typed_segment)>, ReferenceSegment>) {
+          if constexpr (std::is_same_v<IterableType, ReferenceSegmentIterable<T>>) {
             ++reference_chunk_offset;
           }
-        });
+        }
       });
 
       if constexpr (std::is_same_v<Partition<T>, uninitialized_vector<PartitionedElement<T>>>) {  // NOLINT
@@ -184,6 +188,7 @@ RadixContainer<T> materialize_input(const std::shared_ptr<const Table>& in_table
       }
 
       histograms[chunk_id] = std::move(histogram);
+
     }));
     jobs.back()->schedule();
   }