--- a/src/lib/operators/join_index.cpp
+++ b/src/lib/operators/join_index.cpp
@@ -11,7 +11,7 @@
 #include "all_type_variant.hpp"
 #include "join_nested_loop.hpp"
 #include "resolve_type.hpp"
-#include "storage/create_iterable_from_segment.hpp"
+#include "storage/segment_iteration.hpp"
 #include "storage/index/base_index.hpp"
 #include "type_comparison.hpp"
 #include "utils/assert.hpp"
@@ -94,15 +94,8 @@ void JoinIndex::_perform_join() {
       for (ChunkID chunk_id_left = ChunkID{0}; chunk_id_left < input_table_left()->chunk_count(); ++chunk_id_left) {
         const auto segment_left = input_table_left()->get_chunk(chunk_id_left)->get_segment(_column_ids.first);
 
-        resolve_data_and_segment_type(*segment_left, [&](auto left_type, auto& typed_left_segment) {
-          using LeftType = typename decltype(left_type)::type;
-
-          auto iterable_left = create_iterable_from_segment<LeftType>(typed_left_segment);
-
-          // utilize index for join
-          iterable_left.with_iterators([&](auto left_it, auto left_end) {
-            _join_two_segments_using_index(left_it, left_end, chunk_id_left, chunk_id_right, index);
-          });
+        segment_with_iterators_and_data_type_resolve(*segment_left, [&](auto it, const auto end) {
+          _join_two_segments_using_index(it, end, chunk_id_left, chunk_id_right, index);
         });
       }
       performance_data.chunks_scanned_with_index++;