Fix #1145

This introduces `segment_iterate()` and `segment_with_iterators()` as your new way to go to iterate over a segment's data. They compile way faster in debug builds than anything we had previously, as the passed functor is only instantiated once per `DataType`. In release builds, everything remains roughly the same.

I changed the JoinNestedLoop, which took ~15 Minutes to compile in release` to retrieve the outer value via a virtual function call, which greatly reduces its compile time.


ClangDebug libhyrise objects (accumulated):

```
ObjectCount 258 -> 260
BuildTime 632.07s -> 545.41s (-86.66s -13.71%)
FileSize 937.2MiB -> 645.2MiB (-292.0MiB -31.16%)
```

ClangRelease libhyrise objects (accumulated):

```
ObjectCount 258 -> 260
BuildTime 2505.70s -> 1579.45s (-926.25s -36.97%)
FileSize 106.7MiB -> 88.2MiB (-18.5MiB -17.35%)
```