@mrzzzrm 2018-12-03
Apparently there is a performance regression in the JoinHash, so WIP this again.