@mrzzzrm 2018-12-03
Performance regression fixed, `it++` is dangerous.

```
+-----------+-----------------+------+-----------------+------+--------+---------------------------------+
| Benchmark | prev. iter/s    | runs | new iter/s      | runs | change | p-value (significant if <0.001) |
+-----------+-----------------+------+-----------------+------+--------+---------------------------------+
| TPC-H 1   | 2.32604908943   | 140  | 2.35022330284   | 142  | +1%    |     (run time too short) 0.0211 |
| TPC-H 2   | 38.4405822754   | 2307 | 38.5891456604   | 2316 | +0%    |     (run time too short) 0.0000 |
| TPC-H 3   | 20.9787883759   | 1259 | 20.7794380188   | 1247 | -1%    |     (run time too short) 0.0000 |
| TPC-H 4   | 8.18517112732   | 492  | 8.67770385742   | 521  | +6%    |     (run time too short) 0.0000 |
| TPC-H 5   | 25.6755504608   | 1541 | 25.6734676361   | 1541 | -0%    |     (run time too short) 0.9229 |
| TPC-H 6   | 53.4845466614   | 3210 | 53.6534576416   | 3220 | +0%    |     (run time too short) 0.0028 |
| TPC-H 7   | 8.20349502563   | 493  | 8.24221611023   | 495  | +0%    |     (run time too short) 0.9104 |
| TPC-H 8   | 47.3035392761   | 2839 | 47.2677841187   | 2837 | -0%    |     (run time too short) 0.7074 |
| TPC-H 9   | 12.3693571091   | 743  | 12.5635232925   | 754  | +2%    |     (run time too short) 0.0000 |
| TPC-H 10  | 9.75266170502   | 586  | 9.68856334686   | 582  | -1%    |     (run time too short) 0.0000 |
| TPC-H 11  | 133.241668701   | 7995 | 134.474884033   | 8069 | +1%    |     (run time too short) 0.0000 |
| TPC-H 12  | 16.3191566467   | 980  | 16.3221416473   | 980  | +0%    |     (run time too short) 0.9078 |
| TPC-H 13  | 16.358877182    | 982  | 16.2909431458   | 978  | -0%    |     (run time too short) 0.0002 |
| TPC-H 14  | 33.1488876343   | 1989 | 33.1110305786   | 1987 | -0%    |     (run time too short) 0.1860 |
| TPC-H 15  | 13.7051477432   | 823  | 13.7839441299   | 828  | +1%    |     (run time too short) 0.0000 |
| TPC-H 16  | 29.5032062531   | 1771 | 29.4258079529   | 1766 | -0%    |     (run time too short) 0.0000 |
| TPC-H 17  | 1.0299077034    | 62   | 1.02835857868   | 62   | -0%    |                          0.0178 |
| TPC-H 18  | 9.68880081177   | 582  | 9.68684101105   | 582  | -0%    |                          0.8580 |
| TPC-H 19  | 15.5609865189   | 934  | 15.4635782242   | 928  | -1%    |     (run time too short) 0.0000 |
| TPC-H 20  | 0.69080722332   | 42   | 0.705316603184  | 43   | +2%    |                          0.0000 |
| TPC-H 21  | 0.0643693879247 | 4    | 0.0638498663902 | 4    | -1%    |        (not enough runs) 0.0000 |
| TPC-H 22  | 72.7827529907   | 4367 | 71.9938278198   | 4320 | -1%    |     (run time too short) 0.0000 |
| average   |                 |      |                 |      | +0%    |                                 |
+-----------+-----------------+------+-----------------+------+--------+---------------------------------+
```