@mrks 2018-12-04
> it++ is dangerous.

Could you give an example?