@mrzzzrm 2018-12-10
This thing in join_hash_steps:

```c++
        while (it != end) {
          const auto& value = *it;
          ++it;
```

I did this as 


```c++
        while (it != end) {
          const auto& value = *(it++);
```

which created an unnecessary copy of the iterator and took a good chunk out of the JoinHash performance.